chatbot-backend/requirements.txt
fastapi==0.95.0
uvicorn==0.23.1
pydantic==1.10.2
pyspellchecker==0.6.0
python-dotenv==0.21.0
requests==2.28.1
pytest==7.4.0
httpx==0.24.1
starlette>=0.26.1,<0.27.0
pytest-asyncio==0.21.0
python-multipart==0.0.6
pytest-cov==4.0.0
black==23.9.1
isort==5.12.0
mypy==1.5.1
jellyfish==1.0.0
scikit-learn==1.3.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
